requirements_test.txt
pytest-homeassistant-custom-component>=0.13.260
pytest>=8.0
pytest-asyncio>=0.24.0
